requirements.txt
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
google-api-core==2.8.2
google-auth==2.8.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
idna==3.3
numpy==1.23.0
pandas==1.4.3
protobuf==4.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
rsa==4.8
six==1.16.0
toml==0.10.2
urllib3==1.26.9
